Created Scriptable Object Event System.
Created a new Scriptable Object Event System.
- It allows you to create assets inside Unity to use as Events/Listeners
-Event types go from void to Counting Entities (kill/gather), Time (Wait until x or do until x) as well as Defend (health value is less or equal to 0 or health value is higher than 0).
-The system is flexible enough to be integrated into the other systems. One example of this is the Interaction system doing a Count Entity event by calling the Event Asset directly.

The plan is to fully replace the messaging system and use this Scriptable Object one as one can create multiple types of Event Assets to use in various contexts. Furthermore, similar logic is used to assign and/or remove listeners from the event asset itself, meaning that the upgrade process should take minimal effort (minus the sheer amount of scripts to update).